# v2/openapi/fb-unified-openapi.yaml
openapi: 3.0.0
servers:
  - description: Local server
    url: http://0.0.0.0:8000
info:
  title: Fireblocks Provider Connectivity API v2
  version: 0.5.0
  x-logo:
    url: doc-assets/fireblocks-logo-lockup-network-navy.png
    altText: Fireblocks logo
  contact:
    name: Fireblocks
    url: https://fireblocks.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  description: 
    $ref: README.md
...
paths:
...
  /accounts:
    get:
      operationId: getAccounts
      summary: Get list of sub-accounts
      description: Retrieves a paginated list of all sub-accounts available to the provider. Can optionally include balance information for each account if requested. Allows defining parent-child relationships between accounts. Notice that Fireblocks currently supports only one level of hierarchy.
      tags:
        - accounts
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationStartingAfter'
        - $ref: '#/components/parameters/PaginationEndingBefore'
        - $ref: '#/components/parameters/AccountBalancesQueryParam'
      responses:
        '200':
          description: List of sub-accounts.
          content:
            application/json:
              schema:
                type: object
                required:
                  - accounts
                properties:
                  accounts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
  /accounts/{accountId}:
    get:
      operationId: getAccountDetails
      summary: Get sub-account details
      description: |
        Retrieves detailed information about a specific sub-account, including account metadata and optionally balance information if requested.
      tags:
        - accounts
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/AccountBalancesQueryParam'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      responses:
        '200':
          description: List of sub-accounts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
  /accounts/{accountId}/rate:
    get:
      operationId: getRateByAccountAndPairId
      summary: Get rate by account and assets
      tags:
        - rates
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
        - $ref: '#/components/parameters/ConversionPairIdQueryParam'
        - $ref: '#/components/parameters/RampsPairIdQueryParam'
        - $ref: '#/components/parameters/OrderBookPairIdQueryParam'
      responses:
        '200':
          description: Rate by pair id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rate'
              examples:
                usdToEur:
                  summary: USD to EUR rate
                  value:
                    rate: '0.9'
                    timestamp: 1546658861000
                    baseAsset:
                      nationalCurrencyCode: USD
                    quoteAsset:
                      nationalCurrencyCode: EUR
                btcToUsd:
                  summary: BTC to USD rate
                  value:
                    rate: '10000'
                    timestamp: 1546658861000
                    baseAsset:
                      cryptocurrencySymbol: BTC
                    quoteAsset:
                      nationalCurrencyCode: USD
                usdcToUsd:
                  summary: USDC to USD rate
                  value:
                    rate: '1'
                    timestamp: 1546658861000
                    baseAsset:
                      assetId: 360de0ad-9ba1-45d5-8074-22453f193d65
                    quoteAsset:
                      nationalCurrencyCode: USD
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  parameters:
...
    X-FBAPI-KEY:
      name: X-FBAPI-KEY
      in: header
      required: true
      description: API authentication key.
      schema:
        type: string
    X-FBAPI-TIMESTAMP:
      name: X-FBAPI-TIMESTAMP
      in: header
      required: true
      description: Request timestamp in milliseconds since Unix epoch.
      schema:
        type: number
    X-FBAPI-NONCE:
      name: X-FBAPI-NONCE
      in: header
      required: true
      description: Unique identifier of the request.
      schema:
        type: string
    X-FBAPI-SIGNATURE:
      name: X-FBAPI-SIGNATURE
      in: header
      required: true
      description: |-
        Request signature using the chosen cryptographic algorithm. The signature is to be calculated on concatenation of the following request fields in the specified order:
        - `X-FBAPI-TIMESTAMP` - `X-FBAPI-NONCE` - HTTP request method in upper case - Endpoint path, including the query parameters - Request body
      schema:
        type: string
    X-FBAPI-INITIATED-BY:
      name: X-FBAPI-INITIATED-BY
      in: header
      required: false
      description: Conditional. This header is provided only when the corresponding capability requirement is enabled.
      schema:
        type: string
...
    SubAccountIdPathParam:
      name: accountId
      in: path
      required: true
      description: Sub-account identifier.
      schema:
        type: string
        pattern: ^[A-Za-z0-9_-]+$
...
    PaginationLimit:
      name: limit
      in: query
      description: Maximum number of returned items.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 10
    PaginationStartingAfter:
      name: startingAfter
      in: query
      description: Object ID. Instructs to return the items immediately following this object and not including it. Cannot be used together with `endingBefore`.
      schema:
        type: string
    PaginationEndingBefore:
      name: endingBefore
      in: query
      description: Object ID. Instructs to return the items immediately preceding this object and not including it. Cannot be used together with `startingAfter`.
      schema:
        type: string
...
    ConversionPairIdQueryParam:
      name: conversionPairId
      in: query
      description: Conversion pair to get the rate for.
      schema:
        type: string
    RampsPairIdQueryParam:
      name: rampsPairId
      in: query
      description: Ramps pair to get the rate for.
      schema:
        type: string
    OrderBookPairIdQueryParam:
      name: orderBookPairId
      in: query
      description: Order book pair to get the rate for.
      schema:
        type: string
    AccountBalancesQueryParam:
      name: balances
      in: query
      description: Flag to include the account balances in the response. Balances are not returned by default for account endpoints.
      schema:
        type: boolean
...
  responses:
    BadRequestResponse:
      description: Request could not be processed due to a client error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestError'
    UnauthorizedResponse:
      description: Request is unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedError'
    ErrorResponse:
      description: Failed to process request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GeneralError'
...
  schemas:
    GeneralError:
      description: Should be returned only if there in no more specific object.
      type: object
      required:
        - message
        - errorType
      properties:
        message:
          description: Description of the error.
          type: string
        errorType:
          enum:
            - internal-error
            - not-found
    BadRequestErrorType:
      type: string
      enum:
        - schema-error
        - schema-property-error
        - bad-request
        - unknown-asset
        - unsupported-conversion
        - quote-not-ready
        - deposit-address-disabled
        - order-not-trading
        - idempotency-key-reuse
        - deposit-address-creation-not-allowed
        - unsupported-transfer-method
        - transfer-destination-not-allowed
        - unsupported-ramp-method
        - unsupported-source-asset
        - unsupported-destination-asset
        - amount-below-minimum
        - pii-missing
        - unsupported-external-source
        - unsupported-region
        - destination-not-whitelisted
    BadRequestError:
      description: |-
        Request could not be processed due to a client error.
        If the error is caused by a missing or wrong property in the request, `errorType` must have the value `schema-property-error` and the properties  `propertyName` and `requestPart` must be specified. For more general schema errors, the error type should be `schema-error` and `requestPart` must be specified. For all the other cases the error type should be `bad-request`.
      type: object
      required:
        - message
        - errorType
      properties:
        message:
          description: Description of the error.
          type: string
        errorType:
          $ref: '#/components/schemas/BadRequestErrorType'
        propertyName:
          description: |-
            Name of property that caused the error. By convention, should always start with a slash ("/"). If the property is nested, the path should be separated by slashes.
            This property is required if the error type is caused by a missing or wrong property in the request.
          type: string
          example: /topLevelProperyName/childPropertyName/thirdLevelPropertyName
        requestPart:
          $ref: '#/components/schemas/RequestPart'
    UnauthorizedError:
      description: Request is unauthorized
      type: object
      required:
        - message
        - errorType
      properties:
        message:
          description: Description of the error.
          type: string
        errorType:
          enum:
            - unauthorized
        propertyName:
          description: Name of property that caused the error.
          enum:
            - X-FBAPI-KEY
        requestPart:
          description: Request part where the error occurred.
          enum:
            - headers
    RequestPart:
      description: Request part where the error occurred.
      enum:
        - params
        - body
        - querystring
        - headers
...
    PositiveAmount:
      type: string
      pattern: ^\d+(\.\d+)?
    AccountStatus:
      type: string
      enum:
        - active
        - inactive
    AccountData:
      type: object
      required:
        - title
        - status
      properties:
        title:
          type: string
        description:
          type: string
        balances:
          $ref: '#/components/schemas/Balances'
        status:
          $ref: '#/components/schemas/AccountStatus'
        parentId:
          type: string
          description: The ID of the parent account. If the field is empty, the account is a main account. Notice that Fireblocks currently supports only one level of hierarchy.
    AccountId:
      type: object
      required:
        - id
      properties:
        id:
          type: string
    Account:
      allOf:
        - $ref: '#/components/schemas/AccountId'
        - $ref: '#/components/schemas/AccountData'
    Balances:
      type: array
      description: Maps balance assets to their balances.
      items:
        anyOf:
          - $ref: '#/components/schemas/AssetCreditBalance'
          - $ref: '#/components/schemas/AssetBalance'
    AssetBalance:
      type: object
      required:
        - id
        - asset
        - availableAmount
      properties:
        id:
          type: string
        asset:
          $ref: '#/components/schemas/AssetReference'
        availableAmount:
          $ref: '#/components/schemas/PositiveAmount'
        lockedAmount:
          $ref: '#/components/schemas/PositiveAmount'
    AssetCreditBalance:
      allOf:
        - $ref: '#/components/schemas/AssetBalance'
        - type: object
          required:
            - creditAmount
          properties:
            creditAmount:
              $ref: '#/components/schemas/PositiveAmount'
...
    NationalCurrency:
      type: object
      required:
        - nationalCurrencyCode
      properties:
        nationalCurrencyCode:
          $ref: '#/components/schemas/NationalCurrencyCode'
        testAsset:
          type: boolean
    NativeCryptocurrency:
      type: object
      required:
        - cryptocurrencySymbol
      properties:
        blockchain:
          $ref: '#/components/schemas/Blockchain'
        cryptocurrencySymbol:
          $ref: '#/components/schemas/CryptocurrencySymbol'
        testAsset:
          type: boolean
    OtherAssetReference:
      type: object
      required:
        - assetId
      properties:
        assetId:
          description: ID of one of the assets returned in get-additional-assets.
          type: string
...
    Rate:
      type: object
      required:
        - rate
        - timestamp
      properties:
        rate:
          type: string
        timestamp:
          type: number
          format: UNIX-timestamp-epoch
          description: In Milliseconds. The timestamp on the creation of the rate.
          example: 1546658861000
    AssetReference:
      oneOf:
        - $ref: '#/components/schemas/NationalCurrency'
        - $ref: '#/components/schemas/NativeCryptocurrency'
        - $ref: '#/components/schemas/OtherAssetReference'
    Blockchain:
      type: string
      enum:
        - Algorand
        - Arbitrum
        - Astar
        - Aurora
        - Avalanche
        - Axelar
        - Base
        - Berachain
        - Bitcoin
        - Bitcoin Cash
        - Bitcoin Cash ABC
        - Bitcoin SV
        - Bitkub Chain
        - BNB
        - Canto
        - Cardano
        - Celestia
        - Celo
        - Chiliz Chain
        - Chiliz Legacy Chain
        - Cortex
        - Cosmos
        - Dash
        - Dogecoin
        - DYDX
        - eCash
        - EOS
        - Ethereum
        - Ethereum Classic
        - EthereumPoW
        - Evmos
        - Fantom Opera
        - Flare
        - Hedera
        - HT Chain
        - KAVA
        - Kusama
        - Litecoin
        - Moonbeam
        - Moonriver
        - NEAR Protocol
        - NEM NIS1
        - Oasys
        - Optimism
        - Osmosis
        - Palm
        - Polkadot
        - Polygon PoS
        - Ronin
        - RSK
        - Shimmer
        - Solana
        - Songbird Canary-Network
        - Stellar
        - SX Network
        - Terra 2.0
        - Terra Classic
        - Tezos
        - TON
        - TokenEX
        - TRON
        - Velas
        - WEMIX
        - XDB Chain
        - XinFin Network
        - XRP
        - ZCash
    CryptocurrencySymbol:
      type: string
      enum:
        - ADA
        - ALGO
        - ARB
        - ASTR
        - ATOM
        - AURORA
        - AVAX
        - AXL
        - BASE
        - BCH
        - BCHA
        - BERA
        - BITKUB
        - BNB
        - BSV
        - BTC
        - CANTO
        - CELESTIA
        - CELO
        - CHZ
        - CHZ2
        - CTXC
        - DASH
        - DOGE
        - DOT
        - DYDX
        - EOS
        - ETC
        - ETH
        - ETHW
        - EVMOS
        - FLR
        - FTM
        - GLMR
        - HBAR
        - HT
        - KAVA
        - KSM
        - LTC
        - LUNA
        - LUNA2
        - MATIC
        - MOVR
        - NEAR
        - OAS
        - OPT
        - OSMO
        - PALM
        - RON
        - RSK
        - SMR
        - SOL
        - SONGBIRD
        - SXNETWORK
        - TKX
        - TRX
        - VLX
        - WEMIX
        - XDB
        - XDC
        - XEC
        - XEM
        - XLM
        - XRP
        - XTZ
        - ZEC
    NationalCurrencyCode:
      type: string
      description: National currency codes as defined in ISO 4217.
      enum:
        - ADP
        - AED
        - AFA
        - AFN
        - ALK
        - ALL
        - AMD
        - ANG
        - AOA
        - AOK
        - AON
        - AOR
        - ARA
        - ARP
        - ARS
        - ARY
        - ATS
        - AUD
        - AWG
        - AYM
        - AZM
        - BAD
        - BAM
        - BBD
        - BDT
        - BEC
        - BEF
        - BEL
        - BGJ
        - BGK
        - BGL
        - BGN
        - BHD
        - BIF
        - BMD
        - BND
        - BOB
        - BOP
        - BOV
        - BRB
        - BRC
        - BRE
        - BRL
        - BRN
        - BRR
        - BSD
        - BTN
        - BUK
        - BWP
        - BYB
        - BYN
        - BYR
        - BZD
        - CAD
        - CDF
        - CHC
        - CHE
        - CHF
        - CHW
        - CLF
        - CLP
        - CNY
        - COP
        - COU
        - CRC
        - CSD
        - CSJ
        - CSK
        - CUC
        - CUP
        - CVE
        - CYP
        - CZK
        - DDM
        - DEM
        - DJF
        - DKK
        - DOP
        - DZD
        - ECS
        - ECV
        - EEK
        - EGP
        - ERN
        - ESA
        - ESB
        - ESP
        - ETB
        - EUR
        - FIM
        - FJD
        - FKP
        - FRF
        - GBP
        - GEK
        - GEL
        - GHC
        - GHP
        - GHS
        - GIP
        - GMD
        - GNE
        - GNF
        - GNS
        - GQE
        - GRD
        - GTQ
        - GWE
        - GWP
        - GYD
        - HKD
        - HNL
        - HRD
        - HRK
        - HTG
        - HUF
        - IDR
        - IEP
        - ILP
        - ILR
        - ILS
        - INR
        - IQD
        - IRR
        - ISJ
        - ISK
        - ITL
        - JMD
        - JOD
        - JPY
        - KES
        - KGS
        - KHR
        - KMF
        - KPW
        - KRW
        - KWD
        - KYD
        - KZT
        - LAJ
        - LAK
        - LBP
        - LKR
        - LRD
        - LSL
        - LSM
        - LTL
        - LTT
        - LUC
        - LUF
        - LUL
        - LVL
        - LVR
        - LYD
        - MAD
        - MDL
        - MGA
        - MGF
        - MKD
        - MLF
        - MMK
        - MNT
        - MOP
        - MRO
        - MRU
        - MTL
        - MTP
        - MUR
        - MVQ
        - MVR
        - MWK
        - MXN
        - MXP
        - MXV
        - MYR
        - MZE
        - MZM
        - NAD
        - NGN
        - NIC
        - NIO
        - NLG
        - NOK
        - NPR
        - NZD
        - OMR
        - PAB
        - PEH
        - PEI
        - PEN
        - PES
        - PGK
        - PHP
        - PKR
        - PLN
        - PLZ
        - PTE
        - PYG
        - QAR
        - RHD
        - ROK
        - ROL
        - RON
        - RSD
        - RUB
        - RUR
        - RWF
        - SAR
        - SBD
        - SCR
        - SDD
        - SDG
        - SDP
        - SEK
        - SGD
        - SHP
        - SIT
        - SKK
        - SLL
        - SOS
        - SRD
        - SRG
        - SSP
        - STD
        - STN
        - SUR
        - SVC
        - SYP
        - SZL
        - THB
        - TJR
        - TJS
        - TMM
        - TMT
        - TND
        - TOP
        - TPE
        - TRL
        - TRY
        - TTD
        - TWD
        - TZS
        - UAH
        - UAK
        - UGS
        - UGW
        - UGX
        - USD
        - USN
        - USS
        - UYI
        - UYN
        - UYP
        - UYU
        - UYW
        - UZS
        - VEB
        - VEF
        - VES
        - VNC
        - VND
        - VUV
        - WST
        - XAF
        - XCD
        - XOF
        - XPF
        - XSU
        - XTS
        - XUA
        - XXX
        - YDD
        - YER
        - YUD
        - YUM
        - YUN
        - ZAL
        - ZAR
        - ZMK
        - ZMW
        - ZRN
        - ZRZ
        - ZWC
        - ZWD
        - ZWL
        - ZWN
        - ZWR
...
  securitySchemes: {}